requirements.txt
gunicorn==21.2.0
packaging==23.2
django
djangorestframework
channels
channels-redis
redis
daphne
websockets
psycopg==3.1.18
psycopg2==2.9.9
dj-database-url
firebase-admin